Restrict "payment_intent_data" to mode = payment
